Add files via upload
medical_insurance_portfolio.md

## Description

This project does XYZ. It is implemented in Python and is part of my portfolio. 

For this project, I will be investigating a medical insurance costs dataset in a .csv file of 1339 subjects,
using the Python skills that you have developed. 
 
Will Explore areas where the data may include bias and how that would impact potential use cases.
In this code, we analyze the 'smoker' era, and will investigate if and how this status affects the other variables.